Maintenance: upgrade to Reko runtime 0.12.3

+++ parallel/UnitTests/TestArchitecture.cs
@@ -53,7 +53,7 @@ namespace ParallelScan.UnitTests
             throw new NotImplementedException();
         }
 
-        public override FlagGroupStorage GetFlagGroup(RegisterStorage flagRegister, uint grf)
+        public override FlagGroupStorage GetFlagGroup(RegisterStorage flagRegister, ulong grf)
         {
             throw new NotImplementedException();
         }
@@ -88,7 +88,7 @@ namespace ParallelScan.UnitTests
             throw new NotImplementedException();
         }
 
-        public override string GrfToString(RegisterStorage flagRegister, string prefix, uint grf)
+        public override string GrfToString(RegisterStorage flagRegister, string prefix, ulong grf)
         {
             throw new NotImplementedException();
         }
